[Clippy] Address `clippy::unnecessary_cast` lint in src/
In a lot of places, the lint is triggered when casting a constant from the
libgit2 bindings that is created (in libgit2-sys) via `git_enum!`. That macro
will use i32 when `#[cfg(target_env = "msvc")]`, and u32 otherwise, meaning
that a number of the casts that clippy flags as "unnecessary" are indeed
necessary. In those cases, use `#[allow]` with a reason.

Similarly, clippy also lints against casting of `c_int` and `c_uint` to `i32`
and `u32` - while those are usually the same types, the libc documentation
makes it clear that they are not always the same. In those cases too, use
`#[allow]` with a reason.

The only libc-provided type with a consistent rust type is that `size_t` is
always the same as `usize` - fix the places that clippy identified where that
conversion is unneeded. Other fixes include casts when the underlying type is
explicitly set in the libgit2 bindings to be a specific rust type, or when
converting mutable pointers to the same type they already are.

+++ src/cert.rs
@@ -105,6 +105,7 @@ impl<'a> CertHostkey<'a> {
     /// Returns the md5 hash of the hostkey, if available.
     pub fn hash_md5(&self) -> Option<&[u8; 16]> {
         let kind = unsafe { (*self.raw).kind };
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         if kind as u32 & raw::GIT_CERT_SSH_MD5 as u32 == 0 {
             None
         } else {
@@ -115,6 +116,7 @@ impl<'a> CertHostkey<'a> {
     /// Returns the SHA-1 hash of the hostkey, if available.
     pub fn hash_sha1(&self) -> Option<&[u8; 20]> {
         let kind = unsafe { (*self.raw).kind };
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         if kind as u32 & raw::GIT_CERT_SSH_SHA1 as u32 == 0 {
             None
         } else {
@@ -125,6 +127,7 @@ impl<'a> CertHostkey<'a> {
     /// Returns the SHA-256 hash of the hostkey, if available.
     pub fn hash_sha256(&self) -> Option<&[u8; 32]> {
         let kind = unsafe { (*self.raw).kind };
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         if kind as u32 & raw::GIT_CERT_SSH_SHA256 as u32 == 0 {
             None
         } else {
@@ -141,7 +144,7 @@ impl<'a> CertHostkey<'a> {
         Some(unsafe {
             slice::from_raw_parts(
                 (*self.raw).hostkey as *const u8,
-                (*self.raw).hostkey_len as usize,
+                (*self.raw).hostkey_len,
             )
         })
     }
@@ -170,7 +173,7 @@ impl<'a> CertHostkey<'a> {
 impl<'a> CertX509<'a> {
     /// Return the X.509 certificate data as a byte slice
     pub fn data(&self) -> &[u8] {
-        unsafe { slice::from_raw_parts((*self.raw).data as *const u8, (*self.raw).len as usize) }
+        unsafe { slice::from_raw_parts((*self.raw).data as *const u8, (*self.raw).len) }
     }
 }
