+++ src/oid_array.rs
@@ -24,7 +24,7 @@ impl Deref for OidArray {
         unsafe {
             debug_assert_eq!(mem::size_of::<Oid>(), mem::size_of_val(&*self.raw.ids));
 
-            slice::from_raw_parts(self.raw.ids as *const Oid, self.raw.count as usize)
+            slice::from_raw_parts(self.raw.ids as *const Oid, self.raw.count)
         }
     }
 }
